Переименовал файл utils.py в hybrid_cryptosystem.py В этом файле обернул все 3 метода-ситуации в класс, и в main вызываю эти методы через написанный класс.
