+++ lab_3/public_key.pem
@@ -1,9 +1,9 @@
 -----BEGIN PUBLIC KEY-----
-MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAzqx7ZoHUwEK6v0UQESSf
-GqTAuz23cdfFxCYcVLjvoxz6BzVIITaiJN1GwSa/ijk+QrL0O/4V1Qi/tKOfOPKl
-GAWVEmwPBvArX7RA50pQvmYw2ycnCqv6L3Zk2RcYi3pmb8H+vYbUcAmHHlRGNbiA
-WtxtmnUCc8BAzO4ngErsVyk4BRk3sgVOYecubqBLGo0X5GoWmbm9xfdwLwVli10y
-n7TlTWfGcc1+Ea0GxobXrK/B78TB9KsGcYM2S/LEKWGnjYOwdIc5vk3uDW+xOCK3
-laucmzQmJ/eWtj22Rh/mcDEsb1dZOo+EGNCn9X5NepNWTvKPy14iFIt/8bNJlWwi
-BwIDAQAB
+MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAoRtnbE6ugcT24LeO+cbU
+NnbhqkGuCmEPyFp9c6WBMC4jMLGpLXomuo7YDnahBTeti/NOzH9tWz6d8seOHIL6
+J57ZyMS83jMZzDDFc3ropuP68vSsL3WiUXY2egDQ60qOZAtEQR94Mvj8zwVNRFOE
+pJkuou568xHw+CD3knDAVFRCubpmkStFuUbkSDOOMLdxxCbRYOANncCv3UT8BTzA
+QAqKKJR7B3O4k1PpklJfiWBmxaXbBihMnA3Hp3Xq49IUNOcpc09AT6facYNvnS1Y
+qXwC9zwElfbmITqm83jKjW7JOfOpRFAryoYC36HIsehUdx81xuCJxpbVScAxXPov
+zQIDAQAB
 -----END PUBLIC KEY-----
